requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
requests==2.31.0
pydantic==2.6.4
pydantic-settings~=2.2.1
PyYAML==6.0.1
python-dotenv
pdf2image~=1.17.0
tiktoken~=0.9.0
aiohttp~=3.12.2
PyJWT~=2.10.1
cryptography
numpy~=2.2.6
scikit-learn
tenacity>=8.2.3
asyncio
Pillow
python-docx
docx2txt


